Refactored C heliostate, barystate tests.
Simplified the C unit tests for barystate and heliostate
so they use common code for parsing the JPL Horizons
data files. The only thing really different about them
is which function is called to compare against the state
vectors in the files. Pass in a pointer to a context.
Right now the context just contains a pointer to a function
to call, but later it may contain other parameters if needed.