Fixed JavaScript unit tests for Windows.
I noticed on a freshly-cloned Astronomy Engine on Windows,
I started getting JavaScript test failures. The problem was the
script generate/test.js had an inconsistent way of reading text
from files and splitting it into lines. Some of the tests
did not account for \r\n line endings in Windows text files.

So I created a unified approach: a new function ReadLines
that works correctly for Linux and Windows, regardless
of the variant line endings.

This may not work on Mac OS, but I don't have an easy
way of testing that myself. But at least now I can fix
issues like that in one place.