Fixed Windows build issue in makedoc.bat.
This issue affected maintainers (mostly me) only.
Developers who are using Astronomy Engine do not need
to worry about this.

Running the source generator on Windows was broken because
I forgot to copy a step over from the Linux build that
prevents building tutorial pages in the master branch.
I didn't notice this until now because makedoc.bat also
had a bug where it detected an error but did not stop
execution... it just kept going and made it hard to
notice something was wrong.