Raytracer: Explicitly build astronomy.c as C++.
Based on a discovery made by someone helping me get
Astronomy Engine building on the Mac. In some versions
of g++ (aliased to clang), if you don't explicitly tell
the compiler to treat .c files as C++, it gives a warning
that escalates to an error due to -Werror.

So I added the same fix to the raytracer build:
pass in the option '-x c++'.