Browser demos: fix date parsing for Safari.
Safari does not work when you try to do this in JavaScript:

    const date = new Date('2022-01-08 17:26:59');

Instead, it requires a 'T' between the date and the time:

    const date = new Date('2022-01-08T17:26:59');

Fixed this issue in the moonradar.html and positions.html demos.
We still render the date/time as normal, with the space,
but parse it by replacing the space with 'T'.

Thanks to @ebraminio for reporting this!