C moon nodes: confirmed match with Espenak data.
Fixed another bug in my parsing of the original Espenak
data for moon nodes. Added verification that my own
Moon calculations match his:

The Moon is always within 0.182 arcminutes ecliptic
longitude of the node when he says it is crossing the node.

The Moon is always within 1.54 arcminutes of the equatorial
coordinates he says.