Relaxed diff threshold so it passes on my Windows laptop.
The recent changes in C# vector to spherical conversions
caused a miniscule change in the comparison between
C and C# calculations. The same thing happened in Linux.
I am relaxing the threshold slightly so the tests pass
again on my Windows 10 laptop. It's curious that they
didn't fail in the GitHub Actions Windows image.