C unit tests: allow optional tests.
Instead of using conditional compilation to hide
time-consuming performance tests, just mark them
as "optional" so they are only executed when specified
by name, not when we run `ctest all`.

This is a prelude for adding more performance tests,
and is better because it ensures that the code will
build as development continues, even if I'm not using
the test every day.