Added JPL Horizon data for asteroids and comets.
I'm preparing to create a generalized gravity simulator
for small bodies moving through the Solar System
under the influence of the Sun and planets.

Adding JPL Horizons data for 1 comet and 5 asteroids:

    Comet 1P/Halley
    Ceres
    Juno
    Pallas
    Vesta
    101955 Bennu

Halley was chosen as an object with a highly eccentric orbit.

Bennu was chosen as a member of the inner solar system, where
Earth and Mars are likely to have strong effects.

The remaining 4 major asteroids were chosen as typical members
of the asteroid belt between Mars and Jupiter.